pyproject.toml
[tool.poetry]
name = "humob-challenge-2023"
version = "0.1.0"
description = ""
authors = ["osushinekotan"]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
pyyaml = "^6.0.1"
pyarrow = "^12.0.1"
colorlog = "^6.7.0"
torch = "^2.0.1"
plotly = "^5.15.0"
nbformat = "^5.9.2"
dash = "^2.11.1"
ipykernel = "^6.25.0"
jupyter = "^1.0.0"
jupyterlab-code-formatter = "^2.2.1"
jupyterlab = "^4.0.3"
jupyterlab-tabnine = "^0.0.24"
pytorch-pfn-extras = "^0.7.1"
transformers = "^4.31.0"
setuptools = "^68.0.0"
joblib = "^1.3.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.4.1"
ruff = "^0.0.278"
ipykernel = "^6.24.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311" 
unfixable = ["F401", "F841"]
ignore = ["E501"]
line-length = 120

[tool.black]
line-length = 120

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
]
ignore_errors = true